Vision_Based_Safety_Compliance_System/requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
